verify author in `noteController`, user profile route created

+++ src/controller/noteContoller.js
@@ -4,10 +4,12 @@ const Note = require('../models/Note');
 // GET all notes
 const getNotes = async (req, res) => {
 	try {
-		const notes = await Note.find().sort({ updatedAt: -1 });
+		const notes = await Note.find({ author_id: req.user.id }).sort({
+			updatedAt: -1,
+		});
 		res.status(200).json(notes);
-	} catch (err) {
-		return res.status(400).json({ message: err.message });
+	} catch (error) {
+		return res.status(400).json({ message: error.message });
 	}
 };
 
@@ -22,12 +24,16 @@ const getNote = async (req, res) => {
 				.send({ message: `Note with id '${id}' does not exists.` });
 
 		const note = await Note.findById(id);
+
 		if (!note)
 			return res.status(404).send({ message: 'Note does not exists.' });
 
+		if (note.author_id !== req.user.id)
+			return res.status(401).send({ message: 'Note note accessible.' });
+
 		return res.status(200).json(note);
-	} catch (err) {
-		return res.status(400).send({ message: err.message });
+	} catch (error) {
+		return res.status(400).send({ message: error.message });
 	}
 };
 
@@ -39,10 +45,15 @@ const createNote = async (req, res) => {
 		return res.status(400).send({ message: 'Title field cannot be empty.' });
 
 	try {
-		const note = await Note.create({ title, body, style });
+		const note = await Note.create({
+			title,
+			body,
+			style,
+			author_id: req.user.id,
+		});
 		return res.status(200).json(note);
-	} catch (err) {
-		return res.status(400).send({ message: err.message });
+	} catch (error) {
+		return res.status(400).send({ message: error.message });
 	}
 };
 
@@ -51,8 +62,11 @@ const updateNote = async (req, res) => {
 	const { id } = req.params;
 	const { title, body } = req.body;
 
-	if (!title.trim())
-		return res.status(400).send({ message: 'Title field cannot be empty.' });
+	// delete note if title is empty
+	if (!title.trim()) {
+		await Note.deleteOne({ _id: id });
+		return res.status(204);
+	}
 
 	try {
 		if (!mongoose.Types.ObjectId.isValid(id))
@@ -64,11 +78,14 @@ const updateNote = async (req, res) => {
 		if (!note)
 			return res.status(404).send({ message: 'Note does not exists.' });
 
+		if (note.author_id !== req.user.id)
+			return res.status(401).send({ message: 'Note note accessible.' });
+
 		await Note.findOneAndUpdate({ _id: id }, { title, body });
 		const updatedNote = await Note.findById(id);
 		return res.status(200).json(updatedNote);
-	} catch (err) {
-		return res.status(400).send({ message: err.message });
+	} catch (error) {
+		return res.status(400).send({ message: error.message });
 	}
 };
 
@@ -86,10 +103,13 @@ const deleteNote = async (req, res) => {
 		if (!note)
 			return res.status(404).send({ message: 'Note does not exists.' });
 
+		if (note.author_id !== req.user.id)
+			return res.status(401).send({ message: 'Note note accessible.' });
+
 		await Note.deleteOne({ _id: id });
-		return res.status(200).json(note);
-	} catch (err) {
-		return res.status(500).send({ message: err.message });
+		return res.status(204);
+	} catch (error) {
+		return res.status(500).send({ message: error.message });
 	}
 };
